# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Borsa Batoi ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ofertas": {
            "get": {
                "tags": [
                    "ofertas"
                ],
                "summary": "Dades de les ofertes",
                "description": "Torna les dades de les ofertes segons usuari",
                "operationId": "indexOfertas",
                "responses": {
                    "200": {
                        "description": "Ofertes segons el permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OfertaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ofertas"
                ],
                "summary": "Guardar Oferta",
                "description": "Guardar Oferta",
                "operationId": "storeOfertas",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfertaStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Oferta correctament donada d'alta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OfertaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Credencials no vàlides"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/ofertas-arxiu": {
            "get": {
                "tags": [
                    "ofertas"
                ],
                "summary": "Dades de les ofertes",
                "description": "Torna les dades de les ofertes segons usuari",
                "operationId": "indexOfertasArchivadas",
                "responses": {
                    "200": {
                        "description": "Ofertes segons el permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OfertaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/ofertas/{id}": {
            "get": {
                "tags": [
                    "ofertas"
                ],
                "summary": "Dades d'una oferta",
                "description": "Torna les dades d'una oferta amb dades agregades",
                "operationId": "showOfertas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumne segons el permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OfertaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ofertas"
                ],
                "summary": "Modificar Oferta",
                "description": "Modificar Oferta",
                "operationId": "updateOfertas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfertaStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Oferta correctament donada d'alta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OfertaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Credencials no vàlides"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/ofertas/{id}/validar": {
            "put": {
                "tags": [
                    "ofertas"
                ],
                "summary": "Valida Oferta",
                "description": "Torna les dades de l'oferta valida/desvalidada",
                "operationId": "validaOfertas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "validada": {
                                        "title": "validada",
                                        "description": "Oferta Validada",
                                        "type": "boolean",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Oferta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OfertaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/ofertas/{id}/alumno": {
            "put": {
                "tags": [
                    "ofertas"
                ],
                "summary": "Muestra interes por la oferta",
                "description": "Torna les dades de l'oferta",
                "operationId": "InteresOfertas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "interesado": {
                                        "title": "interesado",
                                        "description": "Interes",
                                        "type": "boolean",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Oferta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OfertaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OfertaStoreRequest": {
                "title": "Oferta",
                "description": "Fillable Ofertas",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identificador d'oferta",
                        "type": "integer",
                        "example": "2"
                    },
                    "id_empresa": {
                        "title": "Empresa de la Oferta",
                        "description": "Empresa de l'oferta",
                        "type": "integer",
                        "example": "23"
                    },
                    "descripcion": {
                        "title": "descripcion",
                        "description": "Descripció de l'oferta",
                        "type": "string",
                        "example": "Para arreglar carreteras comarcales"
                    },
                    "puesto": {
                        "title": "puesto",
                        "description": "Puesto de l'oferta",
                        "type": "string",
                        "example": "Peón Caminero"
                    },
                    "tipo_contrato": {
                        "title": "Tipo de contrato",
                        "description": "Tipo de contrato de la oferta",
                        "type": "string",
                        "example": "Indefinido"
                    },
                    "activa": {
                        "title": "activa",
                        "description": "Oferta activa",
                        "type": "boolean",
                        "example": "1"
                    },
                    "contacto": {
                        "title": "contacto",
                        "description": "Persona de contacte de l'empresa",
                        "type": "string",
                        "example": "Pepe Botera"
                    },
                    "telefono": {
                        "title": "telefono",
                        "description": "Telèfon de contacto",
                        "type": "string",
                        "example": "666666666"
                    },
                    "email": {
                        "title": "email",
                        "description": "email de contacte",
                        "type": "string",
                        "example": "pepebotera@gmail.com"
                    },
                    "mostrar_contacto": {
                        "title": "Mostrar contacte",
                        "description": "Mostre contacte de l'empresa",
                        "type": "boolean",
                        "example": "1"
                    },
                    "validada": {
                        "title": "validada",
                        "description": "Oferta Validada",
                        "type": "boolean",
                        "example": "1"
                    },
                    "estudiando": {
                        "title": "estudiando",
                        "description": "Acepta alumnos",
                        "type": "boolean",
                        "example": "1"
                    },
                    "archivada": {
                        "title": "archivada",
                        "description": "Oferta Archivada",
                        "type": "boolean",
                        "example": "0"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Oferta"
                }
            },
...
            "OfertaResource": {
                "title": "Oferta",
                "description": "Ofertas",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identificador de oferta",
                        "type": "integer",
                        "example": "2"
                    },
                    "id_empresa": {
                        "title": "id_empresa",
                        "description": "ID de l'empresa que fa l'oferta",
                        "type": "integer",
                        "example": "5"
                    },
                    "descripcion": {
                        "title": "descripcion",
                        "description": "Descripció de l'oferta",
                        "type": "string",
                        "example": "Se busca persona capaz de aguantar horas de sol"
                    },
                    "puesto": {
                        "title": "puesto",
                        "description": "Lloc de treball",
                        "type": "string",
                        "example": "Peon caminero"
                    },
                    "tipo_contrato": {
                        "schema": "OfertaResource",
                        "title": "tipo_contrato",
                        "description": "Tipo de contrato ofertado",
                        "type": "string",
                        "example": "Temporal"
                    },
                    "activa": {
                        "schema": "OfertaResource",
                        "title": "activa",
                        "description": "Oferta Activa",
                        "type": "boolean",
                        "example": "1"
                    },
                    "contacto": {
                        "schema": "OfertaResource",
                        "title": "contacto",
                        "description": "Persona de contacto",
                        "type": "string",
                        "example": "Pepe Botera"
                    },
                    "telefono": {
                        "schema": "OfertaResource",
                        "title": "telefono",
                        "description": "Telefono de contacto",
                        "type": "string",
                        "example": "546654456"
                    },
                    "email": {
                        "schema": "OfertaResource",
                        "title": "email",
                        "description": "Email de contacto",
                        "type": "string",
                        "example": "PepeBotera@gmail.com"
                    },
                    "mostrar_contacto": {
                        "schema": "OfertaResource",
                        "title": "mostrar_contacto",
                        "description": "Muestro o no el contacto",
                        "type": "boolean",
                        "example": "1"
                    },
                    "validada": {
                        "schema": "OfertaResource",
                        "title": "validada",
                        "description": "Oferta Validada",
                        "type": "boolean",
                        "example": "1"
                    },
                    "estudiando": {
                        "schema": "OfertaResource",
                        "title": "estudiando",
                        "description": "Acepta estudiantes",
                        "type": "boolean",
                        "example": "0"
                    },
                    "interesado": {
                        "schema": "OfertaResource",
                        "title": "interesado",
                        "description": "Si es alumno, si esta interesado en la oferta",
                        "type": "boolean",
                        "example": "0"
                    },
                    "alumnos": {
                        "schema": "OfertaResource",
                        "title": "alumnos",
                        "description": "Si no es alumno, datos alumnos interesado",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlumnoResource"
                        }
                    },
                    "ciclos": {
                        "schema": "OfertaResource",
                        "title": "ciclos requeridos",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "title": "id",
                                    "description": "Identificador de ciclo",
                                    "type": "integer",
                                    "example": "2"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "empresa": {
                        "schema": "OfertaResource",
                        "title": "empresa",
                        "description": "Empresa Oferta",
                        "type": "string",
                        "example": "Pepe Botera y Otilio,S.A"
                    },
                    "created_at": {
                        "schema": "OfertaResource",
                        "title": "created_at",
                        "description": "Hora de creació registre",
                        "type": "time",
                        "example": "2022-03-17 19:37:34"
                    },
                    "updated_at": {
                        "schema": "OfertaResource",
                        "title": "updated_at",
                        "description": "Hora de modificació registre",
                        "type": "time",
                        "example": "2022-03-17 19:37:34"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OfertasResource"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}